Refactoring of reused code in all DetailsMain... components
