+++ src/components/portfolio/DetailsMain.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
-import { makeStyles, createStyles } from '@material-ui/core';
+import { makeStyles, createStyles, Theme } from '@material-ui/core';
+import { useTranslation } from 'react-i18next';
 import DetailsMainSummary from './DetailsMainSummary';
 import DetailsMainPositions from './DetailsMainPositions';
 import DetailsMainRisk from './DetailsMainRisk';
@@ -8,7 +9,7 @@ import DetailsMainDividens from './DetailsMainDividends';
 import DetailsMainAnalyst from './DetailsMainAnalyst';
 import { RiskAnalysis, Position, KeyFigures } from './DetailsTypes';
 
-const useStyles = makeStyles(() =>
+const useStyles = makeStyles(({ palette }: Theme) =>
   createStyles({
     mainWrapper: {
       // TODO use theme margin
@@ -20,8 +21,56 @@ const useStyles = makeStyles(() =>
     sectionWrapper: {
       padding: '2rem 0',
     },
+    titleContainer: {
+      display: 'flex',
+      marginBottom: '2rem',
+    },
+    titleWrapper: {
+      marginRight: '1rem',
+    },
+    sectionTitle: {
+      margin: 0,
+      color: palette.primary.contrastText,
+      // TODO use theme fontsize and weight
+      fontSize: '2.25rem',
+      fontWeight: 400,
+      whiteSpace: 'nowrap',
+    },
+    lineWrapper: {
+      display: 'flex',
+      width: '100%',
+      // TODO: use theme color
+      borderColor: '#EEF1FB',
+    },
+    line: {
+      width: '100%',
+      alignSelf: 'center',
+      paddingLeft: '2%',
+    },
   })
 );
+
+type SectionProps = {
+  title: string;
+};
+
+const Section: React.FC<SectionProps> = ({ title, children }) => {
+  const classes = useStyles();
+
+  return (
+    <div className={classes.sectionWrapper}>
+      <div className={classes.titleContainer}>
+        <div className={classes.titleWrapper}>
+          <h2 className={classes.sectionTitle}>{title}</h2>
+        </div>
+        <div className={classes.lineWrapper}>
+          <hr className={classes.line} />
+        </div>
+      </div>
+      {children}
+    </div>
+  );
+};
 
 type DetailsMainProps = {
   positionCount: number;
@@ -49,10 +98,11 @@ const DetailsMain: React.FC<DetailsMainProps> = ({
   dividendPayoutRatio,
 }) => {
   const classes = useStyles();
+  const { t } = useTranslation();
 
   return (
     <div className={classes.mainWrapper}>
-      <div className={classes.sectionWrapper}>
+      <Section title={t('portfolio.details.summaryHeader')}>
         <DetailsMainSummary
           positionCount={positionCount}
           value={value}
@@ -62,25 +112,25 @@ const DetailsMain: React.FC<DetailsMainProps> = ({
           risk={risk}
           positions={positions}
         />
-      </div>
-      <div className={classes.sectionWrapper}>
+      </Section>
+      <Section title={t('portfolio.details.positionsTitle')}>
         <DetailsMainPositions positions={positions} />
-      </div>
-      <div className={classes.sectionWrapper}>
+      </Section>
+      <Section title={t('portfolio.details.risk')}>
         <DetailsMainRisk risk={risk} positions={positions} />
-      </div>
-      <div className={classes.sectionWrapper}>
+      </Section>
+      <Section title={t('portfolio.details.keyfigures')}>
         <DetailsMainKeyFigures figures={figures} />
-      </div>
-      <div className={classes.sectionWrapper}>
+      </Section>
+      <Section title={t('portfolio.details.dividends')}>
         <DetailsMainDividens
           nextDividend={nextDividend}
           dividendPayoutRatio={dividendPayoutRatio}
         />
-      </div>
-      <div className={classes.sectionWrapper}>
+      </Section>
+      <Section title={t('portfolio.details.analyst')}>
         <DetailsMainAnalyst />
-      </div>
+      </Section>
     </div>
   );
 };
