+++ src/components/portfolio/DetailsMainDividends.tsx
@@ -6,32 +6,6 @@ import RatioDonut from '../charts/RatioDonut';
 // stylesheet for the dividend section
 const useStyles = makeStyles(({ palette }: Theme) =>
   createStyles({
-    titleContainer: {
-      display: 'flex',
-      marginBottom: '2rem',
-    },
-    titleWrapper: {
-      marginRight: '1rem',
-    },
-    sectionTitle: {
-      margin: 0,
-      color: palette.primary.contrastText,
-      // TODO use theme fontsize and weight
-      fontSize: '2.25rem',
-      fontWeight: 400,
-      whiteSpace: 'nowrap',
-    },
-    lineWrapper: {
-      display: 'flex',
-      width: '100%',
-      // TODO: use theme color
-      borderColor: '#EEF1FB',
-    },
-    line: {
-      width: '100%',
-      alignSelf: 'center',
-      paddingLeft: '2%',
-    },
     dividendsWrapper: {
       display: 'flex',
       flexDirection: 'row',
@@ -116,36 +90,22 @@ const DetailsMainDividends: React.FC<DetailsMainDividendsProps> = ({
   const { t } = useTranslation();
 
   return (
-    <div>
-      <div className={classes.titleContainer}>
-        <div className={classes.titleWrapper}>
-          <h2 className={classes.sectionTitle}>
-            {t('portfolio.details.dividends')}
-          </h2>
-        </div>
-        <div className={classes.lineWrapper}>
-          <hr className={classes.line} />
-        </div>
-      </div>
-      <div className={classes.dividendsWrapper}>
-        <div className={classes.chartContainer}>
-          {/* left side with graph */}
-        </div>
-        <div className={classes.infoContainer}>
-          {/* right side with info */}
-          <InfoBlock
-            title={t('portfolio.details.divYield')}
-            body={<p style={{ margin: 0 }}>tmp</p>}
-          />
-          <InfoBlock
-            title={t('portfolio.details.payout')}
-            body={<RatioDonut ratio={dividendPayoutRatio} />}
-          />
-          <InfoBlock
-            title={t('portfolio.details.nextDate')}
-            body={<p style={{ margin: 0 }}>{nextDividend}</p>}
-          />
-        </div>
+    <div className={classes.dividendsWrapper}>
+      <div className={classes.chartContainer}>{/* left side with graph */}</div>
+      <div className={classes.infoContainer}>
+        {/* right side with info */}
+        <InfoBlock
+          title={t('portfolio.details.divYield')}
+          body={<p style={{ margin: 0 }}>tmp</p>}
+        />
+        <InfoBlock
+          title={t('portfolio.details.payout')}
+          body={<RatioDonut ratio={dividendPayoutRatio} />}
+        />
+        <InfoBlock
+          title={t('portfolio.details.nextDate')}
+          body={<p style={{ margin: 0 }}>{nextDividend}</p>}
+        />
       </div>
     </div>
   );
