+++ src/components/portfolio/DetailsMainKeyFigures.tsx
@@ -1,37 +1,10 @@
 import React from 'react';
-import { makeStyles, createStyles, Theme } from '@material-ui/core/styles';
-import { useTranslation } from 'react-i18next';
+import { makeStyles, createStyles } from '@material-ui/core/styles';
 import { KeyFigures } from './DetailsTypes';
 
 // stylesheet for the key figure section
-const useStyles = makeStyles(({ palette }: Theme) =>
+const useStyles = makeStyles(() =>
   createStyles({
-    titleContainer: {
-      display: 'flex',
-      marginBottom: '2rem',
-    },
-    titleWrapper: {
-      marginRight: '1rem',
-    },
-    sectionTitle: {
-      margin: 0,
-      color: palette.primary.contrastText,
-      // TODO use theme fontsize and weight
-      fontSize: '2.25rem',
-      fontWeight: 400,
-      whiteSpace: 'nowrap',
-    },
-    lineWrapper: {
-      display: 'flex',
-      width: '100%',
-      // TODO: use theme color
-      borderColor: '#EEF1FB',
-    },
-    line: {
-      width: '100%',
-      alignSelf: 'center',
-      paddingLeft: '2%',
-    },
     riskContainer: {
       display: 'flex',
       justifyContent: 'space-between',
@@ -55,21 +28,11 @@ type DetailsMainKeyFiguresProps = {
 // returns the details page header
 const DetailsMainKeyFigures: React.FC<DetailsMainKeyFiguresProps> = () => {
   const classes = useStyles();
-  const { t } = useTranslation();
 
   return (
-    <div>
-      <div className={classes.titleContainer}>
-        <div className={classes.titleWrapper}>
-          <h2 className={classes.sectionTitle}>
-            {t('portfolio.details.keyfigures')}
-          </h2>
-        </div>
-        <div className={classes.lineWrapper}>
-          <hr className={classes.line} />
-        </div>
-      </div>
-      <div className={classes.figureWrapper}>{/* body */}</div>
+    <div className={classes.figureWrapper}>
+      {/* body placeholder */}
+      <div />
     </div>
   );
 };
